Move the main file to the correct directory and adjust the module settings

+++ CalculatorTest/src/main/java/Calculator.java
@@ -2,22 +2,22 @@ import java.util.NoSuchElementException;
 
 public class Calculator {
 
-    public static int multiply(int num1, int num2) {
+    public static double multiply(double num1, double num2) {
         return num1 * num2;
     }
 
-    public static int divide(int num1, int num2) throws ArithmeticException{
+    public static double divide(double num1, double num2) throws ArithmeticException{
         if (num2 == 0) {
             throw new ArithmeticException("Cannot be divided by 0");
         }
         return num1/num2;
     }
 
-    public static int sum(int num1, int num2) {
+    public static double sum(double num1, double num2) {
         return num1 + num2;
     }
 
-    public static int subtract(int num1, int num2) {
+    public static double subtract(double num1, double num2) {
         return num1 - num2;
     }
